// components/sections/delivery/types.ts
export interface DeliveryType {
  icon: string
  title: string
  description: string
}

export interface DeliveryOption {
  title: string
  description: string
  distance: string
  icon: string
}

export interface DeliveryAdvantage {
  icon: string
  title: string
  description: string
}

export interface PaymentOption {
  icon: string
  title: string
  description: string
}